requirements.txt
discord.py
praw
python-dotenv